+++ oxana/src/lib.rs
@@ -111,6 +111,7 @@ pub mod prometheus;
 mod test_helper;
 
 pub use crate::context::JobContext;
+pub use crate::drainer::DrainStats;
 pub use crate::error::OxanaError;
 pub use crate::job_envelope::{JobConflictStrategy, JobData, JobEnvelope, JobId, JobMeta};
 pub use crate::job_state::{JobProgress, JobProgressIterator, JobState};
@@ -119,12 +120,15 @@ pub use crate::queue::{
     Queue, QueueConcurrency, QueueConfig, QueueKind, QueueRuntimeConfig, QueueState, QueueThrottle,
     value_to_queue_key,
 };
+pub use crate::result_collector::Stats as RunStats;
 pub use crate::runtime::RuntimeBuilder;
 pub use crate::stats::*;
 pub use crate::storage::Storage;
 pub use crate::storage_builder::{StorageBuilder, StorageBuilderTimeouts};
 pub use crate::storage_types::*;
-pub use crate::worker::{BatchItem, BoxError, FromContext, Job, Worker, WorkerBatchConfig};
+pub use crate::worker::{
+    BatchItem, BoxError, FromContext, IntoWorkerError, Job, Worker, WorkerBatchConfig,
+};
 pub use crate::worker_registry::{
     OnDemandJobRegistration, WorkerConfig, WorkerConfigKind, job_batch_factory,
     job_envelope_factory, job_factory,